Backend/requirements.txt
absl-py==0.12.0
asgiref==3.3.1
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.4
flatbuffers==1.12
gast==0.3.3
google-api-core==1.26.3
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-cloud-dialogflow==2.1.2
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
joblib==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
nltk==3.5
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
proto-plus==1.18.1
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
regex==2021.3.17
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.2
six==1.15.0
sqlparse==0.4.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
tflearn==0.5.0
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
gunicorn
whitenoise==5.0.1
